fix: type errors, config cleanup, and Next.js prerender handling
